update_engine: Remove usage of non-literal format string.
FakeP2PManagerConfiguration used a private variable to store a
format string that unittests can change. This implies that we use a
non-literal format string with printf.

This patch instead uses {file_id} and {minsize} as markers on that
string that get replaced later to the passed values, avoiding the
non-literal format string. This also drops the exception flag from
the gyp file.

BUG=chromium:394166
TEST=USE="clang asan" FEATURES="test" cros_workon_make update_engine

Change-Id: I071a743e756c6214f4915b094520ef73ae902362
Reviewed-on: https://chromium-review.googlesource.com/208205
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>