paycheck: Fix printing of operation index while tracing.
The block tracer is meant to scan operations in reverse order, to
discover the latest operation that writes to a block. Strangely, it only
reversed the operation indexes but scanned the actual operations in the
original order, which is both incorrect in the general case, but even
when it works the printed results are confusing (operations shown with
the wrong index). This fixes it.

Also some cosmetic changes to pacify the linter.

BUG=chromium:510909
TEST=paycheck -B now prints the correct operation indexes.

Change-Id: I65c44eeb450c229a2d5251737a0953716e124687
Reviewed-on: https://chromium-review.googlesource.com/286220
Commit-Queue: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>