update_engine: Set a sane chunk_size default.
When generating deltas of normal ChromeOS images, the expected maximum
size of a file is about 120 MiB, which is reasonble to handle in
memory during payload generation and application. Nevertheless, if we
find a very big file (or pseudo-file) we could easilly run out of
memory when trying to produce a BSDIFF for it. This patch limits
the chunk_size to 200 MiB, which should not impact any normal
ChromeOS payload.

BUG=chromium:485397
TEST=cros_generate_update_payload produces the same payload.

Change-Id: I1c43e619f424253cab20f0d384c7e90d10bd707d
Reviewed-on: https://chromium-review.googlesource.com/286535
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>