update_engine: Fix flaky DeltaPerformerTest.
The update process is required to write the contents of the new
filesystem data in the new partition, but doesn't require to update the
blocks past the "filesystem size". The filesystem size is indeed the
size verity verifies and any data past that point is not verified.

This patch fixes the DeltaPerformerTest to verify the resulting image
up to the filesystem size and not the partition size. This prevents
failures where the contents of the image past the filesystem size are
modified and not restored to the value the test arbitrarily expected.

BUG=None
TEST=Ran DeltaPerformerTests 50 times. This was previously failing.

Change-Id: I3ff9e8b82428bfa752765bc595adf2610f866212
Reviewed-on: https://chromium-review.googlesource.com/287634
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>