update_engine: Upload sample images for testing.
This patch includes the sample images as a compressed .tar.bz2 file in
the source code. The images are generated with the generate_images.sh
script, but the output is also included here. The script requires
root access to create, mount and populate the images with the different
files, which is not reliable enough to run in the builders. Instead,
the images are unpacked from tha .tar.bz2 file when building the
unittests.

These images are not expected to change frequently and the current size
of the .tar.bz2 file is about 3 kB.

BUG=chromium:503145
TEST=`FEATURES=test cros_workon_make update_engine` unpacks the files only when the .tar.bz2 is updated.
TEST=Unit tests still pass.

Change-Id: Ie2be2afdf7a11a7d04b0477767d00b1daa99f1b0
Reviewed-on: https://chromium-review.googlesource.com/292949
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>