Build update_engine from the new directory path.
The update_engine ebuild recreates the old directory structure placing
update_engine as a subdirectory of platform2 in the build environment.
This structure doesn't match the current location of the code and
breaks incremental builds creating symlinks to update_engine in
platform2/ and even in the new update_engine directory.

This patch modifies the .gyp file to reference platform2 sources via
the platform2_root gyp variable and builds update_engine from the new
path. This fixes incremental builds.

For this change to take effect in Chrome OS, the ebuild file is updated
in a separated CL.

BUG=None
TEST=emerge-link update_engine; cros_workon_make update_engine

Change-Id: Icedd344df5c36b0d6ce07bb5c7714eed4e87e590