Remove Chrome OS specific verity_utils module.
The size of the updated region in the partition should match the size
the verified boot chain requires, or at least be bigger. In Chrome OS,
the kernel partition includes a command line with the size of the
rootfs verity will verify, but the rootfs partition is significantly
bigger (about 800 MB more) so we only update what is required.

In the more general Brillo case, the size of the updated area depends
on how verity is used. This patch removes this logic from the
delta_generator and uses the size of the filesystem, which should
match the size of the verity region in Chrome OS devices using ext2.

Bug: 23084776
Test: sudo emerge_update; `mma` on AOSP

Change-Id: I37f341263d86f127d937b4afc24e9d085131cb08