Allow update_engine to create executable files.
update_engine creates directories in some cases, such as the "prefs"
directory. With this mask set, the directories created in Android don't
have the executable bit for the owner, so you can't create files inside
those directories. In Chrome OS this is not really a problem because
update_engine runs as root anyway.

This patch removes the executable bit mask for the owner.

Bug: 23186405
Test: Deployed on drangonboard; /data/misc/update_engine/prefs is executable.

Change-Id: I0c3cc8dd5cacae538091d1f8e131f8247cbf8fbf