Include post-install path in the payload format.
The post-install script was hard-coded to be in /postinst in the root
of the new filesystem. This patch adds support for a configurable path
in the payload to make it easier for the build system to select a
different file in a more standard location, like /bin or /usr/bin.

Note that the PartitionUpdate protobuf message was not used yet by any
version, so we can still make backwards-incompatible changes to it,
such as the field number reorder.

Bug: 24194445
Test: emerge-link update_engine

Change-Id: I4a8c743d21e7d78171a6b825b8e33fb445615279