Read official-build and dev-mode from the system properties.
This patch determines whether we are running in "dev-mode" and whehter
this is an official build in the "ro.secure" and "ro.debuggable" system
properties.

To reduce excessive logging output, this patch moved the logging of
dev-mode and unofficial builds to the beginning of the program and logs
just once.

Bug: 24077521
Test: Deployed an -eng build, update_engine thinks it is in dev-mode and non-official build.

Change-Id: Ib44823cb27dec3d39e1736d982869742b8211a57