Android.mk: Sort library dependencies.
The order of the library dependencies in the makefile is important,
since the dependencies should be listed after the libraries using them
are listed.

This patch sorts the libraries in the Android.mk to match the order in
which their equivalent library dependencies are included in the .gyp
file. It also completes the transitive dependencies between targets as
described by the exported deps in the .gyp file.

Bug: None
Test: mmma system/update_engine

Change-Id: Iaa5bd471919902de9c2646502ff9f918bc9a144a