Disable deadline file in Android.
The deadline file is used to notify chrome the deadline string sent by
the server. This message allows chrome to take a different action based
on the urgency of the update. While we would like to provide a proper
API for this mechanism, it is rarely used in Chrome OS and doesn't work
on Android, so we disable it for now.

Bug: 24868621
Test: Unittest pass. No error message when tested on a dragonboard.

Change-Id: I77fc5b55e1c6ec79f12d88be40426058b8aa4d02