Poll Omaha every 5 hours on Brillo builds.
This simple patch pings Omaha every 5 hours in Brillo instead of the 45
minutes we use in Chrome OS.

All Omaha clients except for Chrome OS, ping every 5 hours so we want
Brillo to do the same as all other devices. For Chrome OS, being offline
is more common so we want to check more often.

Bug: 24341710
Test: Deployed on a device. Policy manager logs says it will check in 5h.

Change-Id: Idc363080f6f60274d66cd372a420be9ac8b4ba08