Android.mk: Remove makefile macro and let make resolve transitive dependencies.
The automatic tool to translate android makefiles to the new build
system doesn't handle macros so we remove them here and follow the
Android style of declaring local make variables.

Because of this change, instead of resolving the transitive dependencies
on each gyp target, we create local variables with the "exported" build
settings and include them from the depending targets. This maps more
closely the gyp targets to the Android.mk targets and allow us to make
conditional changes (based on global make variables) to these variables
in a single place. This will be used to compile weave support
conditionally.

Bug: 24386768
Test: mmma system/update_engine

Change-Id: I9cb44b0e1d4e9f7c061e1b1f154d51c04375aaa0