Send "success reboot" only after an update.
The <event> message sent to omaha after we reboot into a new version
was originally sent only on the first updatecheck after rebooting into
the new installed version, and included the old version number. A bug
introduced in 2013 removed the check for the empty previous version
making the update_engine send this message on *every* updatecheck. This
patch inserts back the check for the empty previous version to send the
"success reboot" event only after.

While there are still discussions about what's the right event message
to send on that case, this fixes the regression.

Bug: 24867646
Test: FEATURES=test emerge-link update_engine; second <updatecheck> didn't include the <event>

Change-Id: Icd00364aae15c8bf420f1f89ae7b526999b3a124