Increase exponential backoff maximum time to 26h on Brillo.
Brillo devices ping Omaha every 5 hours and will increase that value on
an exponential backoff schedule when receiving an error 500 or 503. For
Chrome OS devices, the ping happens every 45 minutes, so the limit on
the exponential backoff is set to 4 hours. This CL changes the limit on
Brillo to 26h, to maintain the proportion.

Bug: 24878416
Test: mm

Change-Id: Ie5ffe6e08a25c2c801ed1aa3b79cb6742cd07caa